Add create, get and delete Customer application
