+++ src/Orderly.Application/Command/Customer/CreateCustomer/CreateCustomerOutput.cs
@@ -1,0 +1,3 @@
+namespace Orderly.Application.Command.Customer.CreateCustomer;
+
+public record CreateCustomerOutput(string CustomerId);
