+++ src/Orderly.Application/Command/Customer/DeleteCustomer/DeleteCustomerInput.cs
@@ -1,0 +1,3 @@
+namespace Orderly.Application.Command.Customer.DeleteCustomer;
+
+public record DeleteCustomerInput(string CustomerId);
