+++ src/Orderly.Application/Command/Customer/DeleteCustomer/DeleteCustomerOutput.cs
@@ -1,0 +1,3 @@
+namespace Orderly.Application.Command.Customer.DeleteCustomer;
+
+public record DeleteCustomerOutput(string CustomerId);
