+++ src/Orderly.Application/Query/Customer/GetCustomerInput.cs
@@ -1,0 +1,3 @@
+namespace Orderly.Application.Query.Customer;
+
+public record GetCustomerInput(string CustomerId);
