--- src/Orderly.Application/UseCase/Customer/GetCustomer/GetCustomerInput.cs
@@ -1,3 +1,0 @@
-namespace Orderly.Application.UseCase.Customer.GetCustomer;
-
-public record GetCustomerInput(string CustomerId);
