+++ tests/UnitTests/Orderly.Application.UnitTests/TestUtils/CreateCustomer/CreateCustomerFixture.cs
@@ -1,6 +1,7 @@
 using Moq;
+using Orderly.Application.Command;
+using Orderly.Application.Command.Customer.CreateCustomer;
 using Orderly.Application.UseCase;
-using Orderly.Application.UseCase.Customer.CreateCustomer;
 using Orderly.Domain.Customer;
 using Orderly.Domain.SalesConsultant;
 using Orderly.Domain.UnitTests.TestUtils.Constants;
